Rename namespace based on folder structure
