+++ src/HE.FMS.Middleware.Common/Exceptions/Internal/MissingConfigurationException.cs
@@ -1,11 +1,3 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using HE.FMS.Middleware.Common.Exceptions.Common;
-using HE.FMS.Middleware.Common.Exceptions.Validation;
-
 namespace HE.FMS.Middleware.Common.Exceptions.Internal;
 public class MissingConfigurationException : InternalSystemException
 {
